Fixed File Extra Dupes Bug

+++ basicwrite.py
@@ -34,7 +34,7 @@ def format_file():
     #skip_copy_directory = ''.join([working_dir, 'Copy_Formatted_CSV_Files','/'])
     #skip_copy_directory_txt = ''.join([working_dir, 'Copy_Formatted_TXT_Files', '/'])
 
-
+    loopcount = int(0)
 
     # ) walks thru every file/dir/root in working directory
 
@@ -44,46 +44,70 @@ def format_file():
 
         for filename in files:  # looks thru every file
 
-            # if not skipdirectory then do the below logic
-            if current_path != wildcardhandler.fnmatch(skip_copy_directory, '*'):  # try and prevent any logic from executing in directory and sub/files not 100% successful
-                print(current_path)
+            print(filename)
+            print(files)
+            print(''.join([os.getcwd(),'/',filename]))
+            test = ''.join([os.getcwd(),'/',filename])
+            print (root)
+            print (dirs)
+            print(os.path.isfile(test))
+            print(os.path.isfile(''.join([working_dir,'/',filename])))
+
+
+            if os.path.isfile(''.join([working_dir,'/',filename])):  # added to fix iteration on directory  #must BE ABSOLUTE PATH
+                print(filename)
+
+                # if not skipdirectory then do the below logic
+                if current_path != wildcardhandler.fnmatch(skip_copy_directory, '*'):  # try and prevent any logic from executing in directory and sub/files not 100% successful
+                    print(current_path)
+
+                    print(loopcount)
+                    loopcount += 1
 
                     #Todo:  try instead of adding it to harcoded path, instead add fnmatch to if logic in order
                     # to skip directory and prevent duplicates
 
-                # check if copy file already exist/ delete if old version/ create new if none
-                exist_dircheck = os.path.isdir(skip_copy_directory)
-                exist_filecheck = os.path.isfile(''.join([skip_copy_directory, filename]))
+                    # check if copy file already exist/ delete if old version/ create new if none
+                    exist_dircheck = os.path.isdir(skip_copy_directory)
+                    exist_filecheck = os.path.isfile(''.join([skip_copy_directory, filename]))
 
 
-                # Make files in new directory, Remove and remake if already there
-                if exist_dircheck:
-                    if exist_filecheck:
+                    # Make files in new directory, Remove and remake if already there(# not the solution)
+                    if exist_dircheck:
+                        if exist_filecheck:
 
-                        try:
-                            os.remove(wildcardhandler.fnmatch(exist_filecheck,'*')) #remove dupes, does not preventtheir creation tho (!!!! failed !!!)
-                        except:
-                            pass
+                            try:
+                                os.remove(wildcardhandler.fnmatch(exist_filecheck,'*')) #remove dupes, does not preventtheir creation tho (!!!! failed !!!)
+                            except:
+                                pass
 
-                        file = open(''.join([skip_copy_directory, filename, '_formatted.csv']), 'w+') # (this is the write file for csv conversion moved inside if logic to fix dup bug
+                            file = open(''.join([skip_copy_directory, filename, '_formatted.csv']), 'w+') # (this is the write file for csv conversion moved inside if logic to fix dup bug
+                        else:
+                            file = open(''.join([skip_copy_directory, filename, '_formatted.csv']),'w+')  # (this is the write file for csv conversion moved inside if logic to fix dup bug
+
                     else:
-                        file = open(''.join([skip_copy_directory, filename, '_formatted.csv']),'w+')  # (this is the write file for csv conversion moved inside if logic to fix dup bug
+                        os.mkdir(skip_copy_directory)
+                        if exist_filecheck:
 
-                else:
-                    os.mkdir(skip_copy_directory)
-                    file = open(''.join([skip_copy_directory, filename, '_formatted.csv']), 'w+') # (this is the write file for csv conversion moved inside if logic to fix
+                            try:
+                                os.remove(wildcardhandler.fnmatch(exist_filecheck,'*')) #remove dupes, does not preventtheir creation tho (!!!! failed !!!)
+                            except:
+                                pass
+
+                            file = open(''.join([skip_copy_directory, filename, '_formatted.csv']), 'w+') # (this is the write file for csv conversion moved inside if logic to fix dup bug
+                        else:
+                            file = open(''.join([skip_copy_directory, filename, '_formatted.csv']),'w+')  # (this is the write file for csv conversion moved inside if logic to fix dup bug
+
+                    # dupe bug still present
+
+                    #Todo Write Text file (create file) code goes here
 
 
-                # dupe bug still present
 
-                #Todo Write Text file (create file) code goes here
-
+                    #)File content formatting and duplications Steps
 
 
-                #)File content formatting and duplications Steps
-
-
-                with open(os.path.join(working_dir,'Test.csv'), 'r', newline='') as file_data: # this is the read file
+                    with open(os.path.join(working_dir,'Test.csv'), 'r', newline='') as file_data: # this is the read file
 
                     #Todo Strip mismatched column data goes here, count max columns,
                     # only write if column data equal to max columns,
@@ -93,31 +117,31 @@ def format_file():
 
 
 
-                    #Convert import file to list to prepare for formatting and copying
-                    listconversion = list(file_data) # filedata converted to list form
-                    split_list = [elements.split(',') for elements in listconversion] # split list into elements
+                        #Convert import file to list to prepare for formatting and copying
+                        listconversion = list(file_data) # filedata converted to list form
+                        split_list = [elements.split(',') for elements in listconversion] # split list into elements
 
-                    # Formatting Outter List(broad, every character in entire list in every pocket)
-                    stripped_list = [re.sub(r'\"', '', str(line)) for line in split_list]  # regex strip quotes from list
-                    stripped_list = [re.sub(r'\'', '', str(line)) for line in stripped_list]  # regex strip newline from list
-                    stripped_list = [re.sub(r'\\n', '', str(line)) for line in stripped_list]  # regex strip newline from list
-                    stripped_list = [re.sub(r'\\r', '', str(line)) for line in stripped_list]  # regex strip newline from list
+                        # Formatting Outter List(broad, every character in entire list in every pocket)
+                        stripped_list = [re.sub(r'\"', '', str(line)) for line in split_list]  # regex strip quotes from list
+                        stripped_list = [re.sub(r'\'', '', str(line)) for line in stripped_list]  # regex strip newline from list
+                        stripped_list = [re.sub(r'\\n', '', str(line)) for line in stripped_list]  # regex strip newline from list
+                        stripped_list = [re.sub(r'\\r', '', str(line)) for line in stripped_list]  # regex strip newline from list
 
 
-                    # go one list deeper
-                    for outer_list_element in stripped_list: # layer 1 deep
+                        # go one list deeper
+                        for outer_list_element in stripped_list: # layer 1 deep
 
-                        #Formatting Inner String(narrow)
-                        outer_list_element = re.sub(r'\[', '', outer_list_element)  # regex strip list marker left from string
-                        outer_list_element = re.sub(r'\]', '', outer_list_element)  # regex strip list marker right from string
-                        outer_list_element = outer_list_element.rstrip() # remove right side end white spaces, or whatever pattern you put into strip function
-                        outer_list_element = outer_list_element.lstrip() # remove left side lead  whitespaces or whatever pattern you put into strip function
-                        outer_list_element = re.sub(r'\,+$', '', outer_list_element)  # regex strip last misc coma
+                            #Formatting Inner String(narrow)
+                            outer_list_element = re.sub(r'\[', '', outer_list_element)  # regex strip list marker left from string
+                            outer_list_element = re.sub(r'\]', '', outer_list_element)  # regex strip list marker right from string
+                            outer_list_element = outer_list_element.rstrip() # remove right side end white spaces, or whatever pattern you put into strip function
+                            outer_list_element = outer_list_element.lstrip() # remove left side lead  whitespaces or whatever pattern you put into strip function
+                            outer_list_element = re.sub(r'\,+$', '', outer_list_element)  # regex strip last misc coma
 
-                        # start appending to blank copy line by line
-                        file.write(outer_list_element)
-                        file.write('\n')
+                            # start appending to blank copy line by line
+                            file.write(outer_list_element)
+                            file.write('\n')
 
-                file.close()
+                    file.close()
 
     return
